// src/api/generated/schemas/updateVendorRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * cid-api
 * LDAP 로그인 + JWT 인가 API
 * OpenAPI spec version: v1
 */

export interface UpdateVendorRequest {
  /**
   * 벤더 명 (UK)
   * @minLength 0
   * @maxLength 100
   */
  vendorNm?: string;
  /**
   * 벤더 타입 (HW/SW/MSP/CSP 등)
   * @minLength 0
   * @maxLength 20
   */
  vendorTpCd?: string;
  /**
   * 담당자 성명
   * @minLength 0
   * @maxLength 50
   */
  chgrNm?: string;
  /**
   * 담당자 이메일
   * @minLength 0
   * @maxLength 100
   */
  chgrEmailAddr?: string;
  /**
   * 담당자 연락처
   * @minLength 0
   * @maxLength 30
   */
  chgrTelNo?: string;
  /**
   * 비고/메모
   * @minLength 0
   * @maxLength 4000
   */
  remk?: string;
  /**
   * 사용 여부 Y/N
   * @minLength 0
   * @maxLength 1
   */
  useYn?: string;
}